Move HostCreator to EndpointCreator
