+++ src/NServiceBus.Core.Tests/Config/When_using_initialization.cs
@@ -14,7 +14,7 @@
 
             endpointConfiguration.TypesToScanInternal(new[] { typeof(FeatureWithInitialization) });
 
-            var ae = Assert.Throws<Exception>(() => HostCreator.BuildEndpointCreator(endpointConfiguration, new ServiceCollection()));
+            var ae = Assert.Throws<Exception>(() => EndpointCreator.Create(endpointConfiguration, new ServiceCollection()));
             var expected = $"Unable to create the type '{nameof(FeatureWithInitialization)}'. Types implementing '{nameof(INeedInitialization)}' must have a public parameterless (default) constructor.";
             Assert.AreEqual(expected, ae.Message);
         }
